# swagger.yaml
openapi: 3.0.3
info:
  title: Event Booking System API
  description: A Mini Event Management System where users can browse events and book tickets.
  version: 1.0.0

servers:
  - url: http://localhost:3000
    description: Local development server

paths:
  /events:
    get:
      summary: List all upcoming events
      description: Retrieves all events that are scheduled in the future.
      operationId: getAllEvents
      tags:
        - Events
      responses:
        '200':
          description: A list of upcoming events.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create a new event
      description: Creates a new event with title, description, date, and total capacity.
      operationId: createEvent
      tags:
        - Events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - date
                - total_capacity
              properties:
                title:
                  type: string
                  example: "Tech Conference 2025"
                description:
                  type: string
                  example: "A conference about the latest in technology."
                date:
                  type: string
                  format: date-time
                  example: "2025-08-15T10:00:00.000Z"
                total_capacity:
                  type: integer
                  minimum: 1
                  example: 100
      responses:
        '201':
          description: Event created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Event created successfully."
                  data:
                    $ref: '#/components/schemas/Event'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /events/{id}/attendance:
    post:
      summary: Record attendance for an event
      description: Takes a booking code as input and returns how many tickets were booked.
      operationId: recordAttendance
      tags:
        - Events
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the event.
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - booking_code
              properties:
                booking_code:
                  type: string
                  format: uuid
                  example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '201':
          description: Attendance recorded successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Attendance recorded successfully."
                  data:
                    type: object
                    properties:
                      event_id:
                        type: integer
                        example: 1
                      booking_code:
                        type: string
                        example: "550e8400-e29b-41d4-a716-446655440000"
                      user_id:
                        type: integer
                        example: 1
                      num_tickets:
                        type: integer
                        example: 2
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Booking not found for this event.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Attendance already recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /bookings:
    post:
      summary: Book tickets for an event
      description: Books ticket(s) for a user. Checks ticket availability and updates remaining_tickets using a database transaction to prevent race conditions. Returns a unique booking code.
      operationId: createBooking
      tags:
        - Bookings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - event_id
              properties:
                user_id:
                  type: integer
                  minimum: 1
                  example: 1
                event_id:
                  type: integer
                  minimum: 1
                  example: 1
                num_tickets:
                  type: integer
                  minimum: 1
                  default: 1
                  example: 2
      responses:
        '201':
          description: Booking created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Booking created successfully."
                  data:
                    type: object
                    properties:
                      booking_id:
                        type: integer
                        example: 1
                      user_id:
                        type: integer
                        example: 1
                      event_id:
                        type: integer
                        example: 1
                      num_tickets:
                        type: integer
                        example: 2
                      booking_code:
                        type: string
                        format: uuid
                        example: "550e8400-e29b-41d4-a716-446655440000"
        '400':
          description: Validation error or not enough tickets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User or event not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /users/{id}/bookings:
    get:
      summary: Get all bookings for a user
      description: Retrieves all bookings made by a specific user along with event details.
      operationId: getUserBookings
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the user.
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: User bookings retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      bookings:
                        type: array
                        items:
                          type: object
                          properties:
                            booking_id:
                              type: integer
                              example: 1
                            booking_code:
                              type: string
                              example: "550e8400-e29b-41d4-a716-446655440000"
                            num_tickets:
                              type: integer
                              example: 2
                            booking_date:
                              type: string
                              format: date-time
                            event_id:
                              type: integer
                              example: 1
                            event_title:
                              type: string
                              example: "Tech Conference 2025"
                            event_description:
                              type: string
                              example: "A conference about the latest in technology."
                            event_date:
                              type: string
                              format: date-time
        '400':
          description: Invalid user ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /users:
    get:
      summary: List all users
      description: Retrieves all registered users.
      operationId: getAllUsers
      tags:
        - Users
      responses:
        '200':
          description: A list of all users.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create a new user
      description: Creates a new user with name and email.
      operationId: createUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
              properties:
                name:
                  type: string
                  example: "John Doe"
                email:
                  type: string
                  format: email
                  example: "john@example.com"
      responses:
        '201':
          description: User created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "User created successfully."
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /users/{id}:
    get:
      summary: Get a single user
      description: Retrieves a user by their ID.
      operationId: getUserById
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the user.
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: User found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid user ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Event:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "Tech Conference 2025"
        description:
          type: string
          example: "A conference about the latest in technology."
        date:
          type: string
          format: date-time
          example: "2025-08-15T10:00:00.000Z"
        total_capacity:
          type: integer
          example: 100
        remaining_tickets:
          type: integer
          example: 98
        created_at:
          type: string
          format: date-time

    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Alice Johnson"
        email:
          type: string
          format: email
          example: "alice@example.com"
        created_at:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "An error message describing what went wrong."